fix: get_block_by_hash endpoint
