+++ crates/core/types/mod.rs
@@ -1,7 +1,6 @@
 mod account;
 mod block;
 mod constants;
-mod engine;
 mod genesis;
 mod receipt;
 mod transaction;
@@ -9,7 +8,6 @@ mod transaction;
 pub use account::*;
 pub use block::*;
 pub use constants::*;
-pub use engine::*;
 pub use genesis::*;
 pub use receipt::*;
 pub use transaction::*;
