+++ crates/rpc/types/mod.rs
@@ -1,2 +1,4 @@
 pub mod block;
+pub mod fork_choice;
+pub mod payload;
 pub mod transaction;
